fix tests, move function to helpers.js, rename function
