+++ test/blocks/chart/chart.test.js
@@ -41,9 +41,12 @@ setConfig(config);
 
 describe('chart', () => {
   let fetch;
+  let paramsGetStub;
 
   before(() => {
     fetch = sinon.stub(window, 'fetch');
+    paramsGetStub = sinon.stub(URLSearchParams.prototype, 'get');
+    paramsGetStub.withArgs('cache').returns('off');
   });
 
   after(() => {
@@ -467,8 +470,6 @@ describe('chart', () => {
   });
 
   it('fetchData functions as expected with cache control enabled', async () => {
-    const paramsGet = sinon.stub(URLSearchParams.prototype, 'get');
-    paramsGet.returns('off');
     const link = document.createElement('a');
     const linkRel = '/drafts/data-viz/line.json';
     link.href = `${linkRel}`;
@@ -476,7 +477,6 @@ describe('chart', () => {
     fetch.withArgs(link.href, { cache: 'reload' }).resolves(goodResponse);
     const response = await fetchData(link);
     expect(response).to.be.true;
-    paramsGet.restore();
   });
 
   it('fetchData returns json given an anchor tag', async () => {
